azure-devops - Moved `getAnnotationValuesFromEntity` to the common package
