+++ workspaces/azure-devops/.changeset/dull-adults-visit.md
@@ -1,0 +1,6 @@
+---
+'@backstage-community/plugin-azure-devops-common': patch
+'@backstage-community/plugin-azure-devops': patch
+---
+
+Moved `getAnnotationValuesFromEntity` to the common package and deprecated the current location. This will be removed in a future version.
