+++ workspaces/azure-devops/plugins/azure-devops-common/package.json
@@ -41,6 +41,7 @@
     "test": "backstage-cli package test"
   },
   "dependencies": {
+    "@backstage/catalog-model": "^1.7.4",
     "@backstage/plugin-catalog-common": "^1.1.4",
     "@backstage/plugin-permission-common": "^0.9.0"
   },
