+++ workspaces/azure-devops/plugins/azure-devops/src/hooks/useBuildRuns.ts
@@ -23,7 +23,7 @@ import {
 import { azureDevOpsApiRef } from '../api';
 import { useApi } from '@backstage/core-plugin-api';
 import useAsync from 'react-use/esm/useAsync';
-import { getAnnotationValuesFromEntity } from '../utils';
+import { getAnnotationValuesFromEntity } from '@backstage-community/plugin-azure-devops-common';
 import { Entity, stringifyEntityRef } from '@backstage/catalog-model';
 
 export function useBuildRuns(
