+++ workspaces/azure-devops/plugins/azure-devops/src/hooks/useReadme.ts
@@ -20,7 +20,7 @@ import { Entity, stringifyEntityRef } from '@backstage/catalog-model';
 import { azureDevOpsApiRef } from '../api';
 import { useApi } from '@backstage/core-plugin-api';
 import useAsync from 'react-use/esm/useAsync';
-import { getAnnotationValuesFromEntity } from '../utils';
+import { getAnnotationValuesFromEntity } from '@backstage-community/plugin-azure-devops-common';
 
 export function useReadme(entity: Entity): {
   item?: Readme;
