+++ workspaces/azure-devops/plugins/azure-devops/src/utils/index.ts
@@ -17,4 +17,3 @@
 export * from './arrayHas';
 export * from './equalsIgnoreCase';
 export * from './getDurationFromDates';
-export * from './getAnnotationValuesFromEntity';
